[WIP] Finite Differencing along coloring vector
This PR aims to tackle to issue of calculating the jacobian
matrix using finite difference method, particularly in specific
chosen directions that are computed via the coloring of the
graph representation of the derivative matrix.

The function to calculate the direction vectors given the
coloring has been added, however, it will be integrated with
the main finite-diff function in the next commit.