fixed broken syntax
The find_edge method has been reverted back to the old version with a worst case running time of O(size of E) where E is the edge set of graph g. The primary reason is that while the function can be optimized by returning a new edge, the ambiguity between which vertex is the source and which one is the destination breaks the disjoint set function as `Edge 1 => 2` is treated differently compared to `Edge 2=>1`.